Extensive changes made after the RC1 release 02.04.2003. Re-considered auto-setup of aqueous and gas/fluid phases. Added a new AutoPhaseDialog. Created an extended set of built-in functions for calculating activity coeffs for aqueous species, as well as simple solid solution models.
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@228 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/s_fgl.h
@@ -381,7 +381,7 @@ public:
    double DENSITY(double *X,float *param, unsigned NN ,double Pbar, double T );
    int CGActivCoefRhoT(double *X,float *param, double *act, unsigned NN,
      double ro, double T );
-   int CGActivCoefPT(double *X,float *param,double *act, unsigned NN,
+   double CGActivCoefPT(double *X,float *param,double *act, unsigned NN,
      double Pbar, double T );
 
     int CGcalcFug( void );  // Calc. fugacity for 1 species at X=1
